rename misspelled riccati stuff. changing test.
