+++ drake/math/BUILD
@@ -17,11 +17,10 @@ drake_cc_library(
     ],
 )
 
-# TODO(jwnimmer-tri) This should be riccati not ricatti.
 drake_cc_library(
-    name = "continuous_algebraic_ricatti_equation",
-    srcs = ["continuous_algebraic_ricatti_equation.cc"],
-    hdrs = ["continuous_algebraic_ricatti_equation.h"],
+    name = "continuous_algebraic_riccati_equation",
+    srcs = ["continuous_algebraic_riccati_equation.cc"],
+    hdrs = ["continuous_algebraic_riccati_equation.h"],
     deps = [
         "//drake/common",
         "//drake/common:is_approx_equal_abstol",
@@ -181,7 +180,7 @@ drake_cc_googletest(
 drake_cc_googletest(
   name = "continuous_algebraic_riccati_equation_test",
   deps = [
-    ":continuous_algebraic_ricatti_equation",
+    ":continuous_algebraic_riccati_equation",
     "//drake/common:eigen_matrix_compare",
   ],
 )
