+++ drake/math/CMakeLists.txt
@@ -3,7 +3,7 @@ set(sources
   autodiff.cc
   autodiff_gradient.cc
   axis_angle.cc
-  continuous_algebraic_ricatti_equation.cc
+  continuous_algebraic_riccati_equation.cc
   cross_product.cc
   discrete_algebraic_riccati_equation.cc
   eigen_sparse_triplet.cc
@@ -25,7 +25,7 @@ set(installed_headers
   autodiff.h
   autodiff_gradient.h
   axis_angle.h
-  continuous_algebraic_ricatti_equation.h
+  continuous_algebraic_riccati_equation.h
   cross_product.h
   discrete_algebraic_riccati_equation.h
   eigen_sparse_triplet.h
