+++ drake/math/continuous_algebraic_riccati_equation.cc
@@ -1,4 +1,4 @@
-#include "drake/math/continuous_algebraic_ricatti_equation.h"
+#include "drake/math/continuous_algebraic_riccati_equation.h"
 
 #include "drake/common/drake_assert.h"
 #include "drake/common/is_approx_equal_abstol.h"
