+++ drake/math/continuous_algebraic_riccati_equation.h
